proposed edits to README
The previous readme was too presumptuous. :)
Everybody knows that the simplest way to get a command line interface for an OCaml program is to use the minicli library.
And yes, minicli does not generate a manpage for you, and yes it takes seconds with minicli
to add a new option or change an existing one, which is just wonderful when prototyping software.
Also, minicli makes it difficult to use a CLI incorrectly. Something which is quite interesting for scientific
software and that I found nowhere else. :)